requirements.txt
torchdata
diffusers
hydra-core
omegaconf
tqdm
accelerate
wandb
einops
transformers
alias_free_torch
h5py
torchaudio
soundfile
tensorboard
swanlab
librosa